Fixes #42: Move internal module to _utils
